requirements.txt
# Flask and Extensions
Flask>=3.0.0
Flask-Admin>=1.6.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-SQLAlchemy>=3.1.1
Flask-Cors>=4.0.0

# Database
SQLAlchemy>=2.0.0

# Security
Werkzeug>=3.0.0
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0
pillow-heif>=0.15.0

# Machine Learning
tensorflow>=2.15.0
numpy>=1.24.0

# Production Server
gunicorn>=21.0.0

# Utilities
requests>=2.31.0
